+++ CHANGES.md
@@ -1,13 +1,14 @@
 # CHANGES between Releases
 
 ## What's new in Version 1.8.1 of RaspiBlitz?
- 
+
 - New: Support of X708 UPS HAT [details](https://github.com/rootzoll/raspiblitz/pull/3087)
 - Update: LND v0.15.2 [details](https://github.com/lightningnetwork/lnd/releases/tag/v0.15.2-beta
 - Update: Core Lightning v0.12.1 [details](https://github.com/ElementsProject/lightning/releases/tag/v0.12.1)
 - Update: RTL v0.13.0 with update option [details](https://github.com/Ride-The-Lightning/RTL/releases/tag/v0.12.3)
 - Update: LNbits 0.9.2 [details](https://github.com/lnbits/lnbits-legend/releases/tag/0.9.2)
 - Update: ItchySats 0.7.0 [details](https://github.com/itchysats/itchysats/releases/tag/0.7.0)
+- Update: Jam (JoinMarket Web UI) v0.1.1 [details](https://github.com/joinmarket-webui/joinmarket-webui/releases/tag/v0.1.0)
 - Info: Run RaspiBlitz on Proxmox [details](https://github.com/rootzoll/raspiblitz/tree/dev/alternative.platforms/Proxmox)
 
 ## What's new in Version 1.8.0 of RaspiBlitz?